Windows 2016 TP4 support
A product key is required to complete the headless install of Win-2016
TP4. During first boot, it displays a ProductKey prompt and if one is
not provided will not start the network services for remote login.
Keys are at https://technet.microsoft.com/en-us/library/mt126134.aspx as
of this commit.